# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Legal CRM API
  version: 0.1.0
servers:
  - url: http://localhost:3000/api
paths:
...
  /workflows/templates/{id}/instantiate:
    post:
      summary: Instantiate template to matter
      tags: [Workflows]
      description: Creates a workflow instance from template for a specific matter
      security:
        - bearerAuth: []
      parameters:
        - in: path
          required: true
          name: id
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [matterId]
              properties:
                matterId:
                  type: string
                  description: ID of the matter to attach workflow to
      responses:
        '201':
          description: Workflow instance created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkflowInstance'
        '400':
          description: Template has no steps
        '404':
          description: Template or matter not found
        '409':
          description: Template is not published
...
  /workflows/instances/{id}/steps:
    post:
      summary: Add step to workflow instance
      tags: [Workflows]
      description: Add a new step to an existing workflow instance
      security:
        - bearerAuth: []
      parameters:
        - in: path
          required: true
          name: id
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WorkflowStepCreate'
      responses:
        '201':
          description: Step added
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkflowInstanceStep'
        '403':
          description: Access denied
        '404':
          description: Instance not found
...
  /workflows/steps/{id}/skip:
    post:
      summary: Skip a workflow step
      tags: [Workflows]
      description: Marks step as SKIPPED (admin only)
      security:
        - bearerAuth: []
      parameters:
        - in: path
          required: true
          name: id
          schema: { type: string }
      responses:
        '200':
          description: Step skipped
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkflowInstanceStep'
        '403':
          description: Admin access required
        '404':
          description: Step not found
components:
  schemas:
...
    WorkflowInstance:
      type: object
      required: [id, templateId, matterId, status, createdAt]
      properties:
        id: { type: string, format: uuid }
        templateId: { type: string, format: uuid }
        matterId:
          type: string
          format: uuid
          description: Matter this workflow is attached to
        templateVersion:
          type: integer
          description: Version of template used
        status:
          type: string
          enum: [DRAFT, ACTIVE, PAUSED, COMPLETED, CANCELED]
          description: Current workflow status
        context:
          type: object
          description: Shared workflow context data
          additionalProperties: true
        createdById:
          type: string
          format: uuid
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
        template:
          type: object
          properties:
            id: { type: string, format: uuid }
            name: { type: string }
        createdBy:
          type: object
          nullable: true
          properties:
            id: { type: string, format: uuid }
            name: { type: string, nullable: true }
            email: { type: string, format: email, nullable: true }
        steps:
          type: array
          items:
            $ref: '#/components/schemas/WorkflowInstanceStep'
    WorkflowInstanceStep:
      type: object
      required: [id, instanceId, order, title, actionType, roleScope, actionState]
      properties:
        id: { type: string, format: uuid }
        instanceId: { type: string, format: uuid }
        templateStepId:
          type: string
          format: uuid
          nullable: true
          description: Reference to template step (if from template)
        order:
          type: integer
          description: Step order in workflow
        title:
          type: string
          example: "Collect Client Signature"
        actionType:
          type: string
          enum: [APPROVAL_LAWYER, SIGNATURE_CLIENT, REQUEST_DOC_CLIENT, PAYMENT_CLIENT, CHECKLIST]
        roleScope:
          type: string
          enum: [ADMIN, LAWYER, PARALEGAL, CLIENT]
        actionState:
          type: string
          enum: [PENDING, READY, IN_PROGRESS, BLOCKED, COMPLETED, FAILED, SKIPPED]
          description: Current state of the step
        actionConfig:
          type: object
          description: Action configuration
          additionalProperties: true
        actionData:
          type: object
          nullable: true
          description: Runtime action data
          additionalProperties: true
        assignedToId:
          type: string
          format: uuid
          nullable: true
          description: User assigned to this step
        startedAt:
          type: string
          format: date-time
          nullable: true
        completedAt:
          type: string
          format: date-time
          nullable: true
        updatedAt: { type: string, format: date-time }
    WorkflowStepCreate:
      type: object
      required: [title, actionType, roleScope]
      properties:
        title:
          type: string
          minLength: 1
          example: "Review Documents"
        actionType:
          type: string
          enum: [APPROVAL_LAWYER, SIGNATURE_CLIENT, REQUEST_DOC_CLIENT, PAYMENT_CLIENT, CHECKLIST]
        roleScope:
          type: string
          enum: [ADMIN, LAWYER, PARALEGAL, CLIENT]
        required:
          type: boolean
          default: true
        order:
          type: integer
          description: Position in workflow (auto-assigned if not provided)
        actionConfig:
          type: object
          description: Action-specific configuration
          additionalProperties: true
          example:
            items:
              - "Verify ID"
              - "Check address"
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT